+++ api/src/extensions/lib/sandbox/generate-host-function-reference.ts
