+++ api/src/extensions/lib/sandbox/register/index.ts
@@ -1,0 +1,5 @@
+export * from './action.js';
+export * from './call-reference.js';
+export * from './filter.js';
+export * from './operation.js';
+export * from './route.js';
