+++ api/src/extensions/lib/sandbox/sdk/generators/request.ts
@@ -1,31 +1,7 @@
 import type { ExtensionSandboxRequestedScopes } from '@directus/extensions';
 import encodeUrl from 'encodeurl';
 import type { Reference } from 'isolated-vm';
-import { setTimeout } from 'node:timers/promises';
-import logger from '../logger.js';
-import { getAxios } from '../request/index.js';
-
-export function logGenerator(_requestedScopes: ExtensionSandboxRequestedScopes): (message: Reference<string>) => void {
-	return (message) => {
-		if (message.typeof !== 'string') throw new Error('Log message has to be of type string');
-
-		const messageCopied = message.copySync();
-
-		logger.info(messageCopied);
-	};
-}
-
-export function sleepGenerator(
-	_requestedScopes: ExtensionSandboxRequestedScopes
-): (milliseconds: Reference<number>) => Promise<void> {
-	return async (milliseconds) => {
-		if (milliseconds.typeof !== 'number') throw new Error('Sleep milliseconds has to be of type number');
-
-		const millisecondsCopied = await milliseconds.copy();
-
-		await setTimeout(millisecondsCopied);
-	};
-}
+import { getAxios } from '../../../../../request/index.js';
 
 export function requestGenerator(requestedScopes: ExtensionSandboxRequestedScopes): (
 	url: Reference<string>,
