+++ api/src/extensions/lib/sandbox/sdk/index.ts
@@ -1,0 +1,2 @@
+export * from './instantiate.js';
+export * from './sdk.js';
