+++ api/src/extensions/lib/sandbox/sdk/utils/wrap.ts
@@ -1,0 +1,9 @@
+export function wrap(util: any) {
+	return async (...args: any[]) => {
+		try {
+			return { result: await util(...args), error: false };
+		} catch (e) {
+			return { result: e, error: true };
+		}
+	};
+}
