+++ api/src/extensions/lib/wrap-embeds.ts
